refactors Recommender and improves response time
